for #1416, refactor visitor package
