+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/ColumnDefinitionVisitor.java
@@ -18,10 +18,10 @@
 package io.shardingsphere.core.parsing.antlr.visitor.phrase;
 
 import com.google.common.base.Optional;
-import io.shardingsphere.core.parsing.antlr.RuleName;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.ColumnDefinition;
-import io.shardingsphere.core.parsing.antlr.util.ASTUtils;
-import io.shardingsphere.core.parsing.antlr.util.VisitorUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.RuleName;
+import io.shardingsphere.core.parsing.antlr.visitor.util.ASTUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.util.VisitorUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import io.shardingsphere.core.parsing.parser.sql.ddl.create.table.CreateTableStatement;
 import io.shardingsphere.core.util.SQLUtil;
