+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/IndexNameVisitor.java
@@ -18,9 +18,9 @@
 package io.shardingsphere.core.parsing.antlr.visitor.phrase;
 
 import com.google.common.base.Optional;
-import io.shardingsphere.core.parsing.antlr.RuleName;
-import io.shardingsphere.core.parsing.antlr.util.ASTUtils;
-import io.shardingsphere.core.parsing.antlr.util.VisitorUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.RuleName;
+import io.shardingsphere.core.parsing.antlr.visitor.util.ASTUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.util.VisitorUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import io.shardingsphere.core.parsing.parser.sql.ddl.DDLStatement;
 import org.antlr.v4.runtime.ParserRuleContext;
