+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/TableNamesVisitor.java
@@ -17,8 +17,8 @@
 
 package io.shardingsphere.core.parsing.antlr.visitor.phrase;
 
-import io.shardingsphere.core.parsing.antlr.RuleName;
-import io.shardingsphere.core.parsing.antlr.util.ASTUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.RuleName;
+import io.shardingsphere.core.parsing.antlr.visitor.util.ASTUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.tree.ParseTree;
