+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/dialect/mysql/MySQLAddColumnVisitor.java
@@ -20,8 +20,8 @@ package io.shardingsphere.core.parsing.antlr.visitor.phrase.dialect.mysql;
 import com.google.common.base.Optional;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.ColumnPosition;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.mysql.MySQLAlterTableStatement;
-import io.shardingsphere.core.parsing.antlr.util.VisitorUtils;
 import io.shardingsphere.core.parsing.antlr.visitor.phrase.AddColumnVisitor;
+import io.shardingsphere.core.parsing.antlr.visitor.util.VisitorUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.tree.ParseTree;
