+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/dialect/mysql/MySQLChangeColumnVisitor.java
@@ -18,13 +18,13 @@
 package io.shardingsphere.core.parsing.antlr.visitor.phrase.dialect.mysql;
 
 import com.google.common.base.Optional;
-import io.shardingsphere.core.parsing.antlr.RuleName;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.ColumnDefinition;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.ColumnPosition;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.mysql.MySQLAlterTableStatement;
-import io.shardingsphere.core.parsing.antlr.util.ASTUtils;
-import io.shardingsphere.core.parsing.antlr.util.VisitorUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.RuleName;
 import io.shardingsphere.core.parsing.antlr.visitor.phrase.PhraseVisitor;
+import io.shardingsphere.core.parsing.antlr.visitor.util.ASTUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.util.VisitorUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import org.antlr.v4.runtime.ParserRuleContext;
