+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/dialect/sqlserver/SQLServerAddPrimaryKeyVisitor.java
@@ -18,11 +18,11 @@
 package io.shardingsphere.core.parsing.antlr.visitor.phrase.dialect.sqlserver;
 
 import com.google.common.base.Optional;
-import io.shardingsphere.core.parsing.antlr.RuleName;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.AlterTableStatement;
 import io.shardingsphere.core.parsing.antlr.sql.ddl.ColumnDefinition;
-import io.shardingsphere.core.parsing.antlr.util.ASTUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.RuleName;
 import io.shardingsphere.core.parsing.antlr.visitor.phrase.PhraseVisitor;
+import io.shardingsphere.core.parsing.antlr.visitor.util.ASTUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.tree.ParseTree;
