+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/visitor/phrase/dialect/sqlserver/SQLServerDropIndexVisitor.java
@@ -18,10 +18,10 @@
 package io.shardingsphere.core.parsing.antlr.visitor.phrase.dialect.sqlserver;
 
 import com.google.common.base.Optional;
-import io.shardingsphere.core.parsing.antlr.RuleName;
-import io.shardingsphere.core.parsing.antlr.util.ASTUtils;
-import io.shardingsphere.core.parsing.antlr.util.VisitorUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.RuleName;
 import io.shardingsphere.core.parsing.antlr.visitor.phrase.PhraseVisitor;
+import io.shardingsphere.core.parsing.antlr.visitor.util.ASTUtils;
+import io.shardingsphere.core.parsing.antlr.visitor.util.VisitorUtils;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import org.antlr.v4.runtime.ParserRuleContext;
